Genesis - pardon the huge commit
Long story short, this project started as two separate repositories
but I decided to try and merge them together for practicity reasons.
I'm not sure if this is a good idea, but I'm also not sure I should be
coding these weird languages for a probably dead-on-arrival project
instead of developing my agile development soft skills so here we are.