De-optimize docker image to be simpler and more useful
The new docker/compose file will generate kind of huge images instead
of the 50mb one it did, but now we can run tests on it and it is much
simpler to read. The idea is that now tests are more valuable than
the optimization. If we really need it in the future, we can have
two images, no big deal.