Move everything around and create files so I feel productive
It's been a long day, but re-shuffling the files makes me feel good
for some reason. I think it's because the compiler clicks it all
together again like it's some sort of code kniting sorcerer person.

Well, all in all this seems more clean and easier to follow than before.
I'll leave it like that for now.